+++ functions/enqueue.php
@@ -55,30 +55,9 @@
 	// WP Footer hook
 	add_action('wp_footer', function() {
 
-		// PrismJS
-		wp_enqueue_style('prism-css');
-
 		// Define ajaxurl
 		if (is_page(PAGE_CONTACT)):
 			?><script type="text/javascript">var ajaxurl ='<?php echo admin_url('admin-ajax.php'); ?>';</script><?php
-		endif;
-
-		// Social media scripts
-		if (is_singular('post')):
-			?>
-			<!-- Facebook -->
-			<div id="fb-root"></div>
-			<script>(function(d, s, id) {
-				var js, fjs = d.getElementsByTagName(s)[0];
-				if (d.getElementById(id)) return;
-				js = d.createElement(s); js.id = id;
-				js.src = "//connect.facebook.net/en_US/sdk.js#xfbml=1&version=v2.8";
-				fjs.parentNode.insertBefore(js, fjs);
-			}(document, 'script', 'facebook-jssdk'));</script>
-
-			<!-- Twitter -->
-			<script async src="//platform.twitter.com/widgets.js" charset="utf-8"></script>
-			<?php
 		endif;
 	});