+++ functions/theme-functions.php
@@ -159,4 +159,24 @@
 		<?php
 	}
 
+	/**
+	 * Get social media data via cURL
+	 * @param  string $url URL to get data from
+	 * @return object      Rretreived data in JSON format
+	 */
+	function getCurlData($url)
+	{
+		$ch = curl_init();
+	    curl_setopt($ch, CURLOPT_URL, $url);
+	    curl_setopt($ch, CURLOPT_HEADER, 0);
+	    curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+
+	    $response = curl_exec($ch);
+	    if($response === false) echo 'Curl error: ' . curl_error($ch);
+	    	else $response = $response;
+
+	    curl_close($ch);
+	    return json_decode($response);
+	}
+
 ?>